+++ crates/aviutl2/src/filter/bridge.rs
@@ -394,8 +394,8 @@ macro_rules! register_filter_plugin {
         ::aviutl2::__internal_module! {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn InitializeLogger(logger: *mut $crate::sys::logger2::LOG_HANDLE) {
-                $crate::__macro_if! {
-                    if unwind in (unwind = true, $( $key = $value ),* ) {
+                $crate::comptime_if::comptime_if! {
+                    if unwind where (unwind = true, $( $key = $value ),* ) {
                         if let Err(panic_info) = $crate::__catch_unwind_with_panic_info(|| {
                             $crate::logger::__initialize_logger(logger)
                         }) {
@@ -414,8 +414,8 @@ macro_rules! register_filter_plugin {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn InitializePlugin(version: u32) -> bool {
                 unsafe {
-                    $crate::__macro_if! {
-                        if unwind in (unwind = true, $( $key = $value ),* ) {
+                    $crate::comptime_if::comptime_if! {
+                        if unwind where (unwind = true, $( $key = $value ),* ) {
                             $crate::filter::__bridge::initialize_plugin_c_unwind::<$struct>(version)
                         } else {
                             $crate::filter::__bridge::initialize_plugin_c::<$struct>(version)
@@ -427,8 +427,8 @@ macro_rules! register_filter_plugin {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn UninitializePlugin() {
                 unsafe {
-                    $crate::__macro_if! {
-                        if unwind in (unwind = true, $( $key = $value ),* ) {
+                    $crate::comptime_if::comptime_if! {
+                        if unwind where (unwind = true, $( $key = $value ),* ) {
                             $crate::filter::__bridge::uninitialize_plugin_c_unwind::<$struct>()
                         } else {
                             $crate::filter::__bridge::uninitialize_plugin::<$struct>()
@@ -440,8 +440,8 @@ macro_rules! register_filter_plugin {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn GetFilterPluginTable()
             -> *mut aviutl2::sys::filter2::FILTER_PLUGIN_TABLE {
-                $crate::__macro_if! {
-                    if unwind in (unwind = true, $( $key = $value ),* ) {
+                $crate::comptime_if::comptime_if! {
+                    if unwind where (unwind = true, $( $key = $value ),* ) {
                         unsafe { $crate::filter::__bridge::create_table_unwind::<$struct>() }
                     } else {
                         unsafe { $crate::filter::__bridge::create_table::<$struct>() }