+++ crates/aviutl2/src/input/bridge.rs
@@ -781,8 +781,8 @@ macro_rules! register_input_plugin {
         ::aviutl2::__internal_module! {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn InitializeLogger(logger: *mut $crate::sys::logger2::LOG_HANDLE) {
-                $crate::__macro_if! {
-                    if unwind in (unwind = true, $( $key = $value ),* ) {
+                $crate::comptime_if::comptime_if! {
+                    if unwind where (unwind = true, $( $key = $value ),* ) {
                         if let Err(panic_info) = $crate::__catch_unwind_with_panic_info(|| {
                             $crate::logger::__initialize_logger(logger)
                         }) {
@@ -801,8 +801,8 @@ macro_rules! register_input_plugin {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn InitializePlugin(version: u32) -> bool {
                 unsafe {
-                    $crate::__macro_if! {
-                        if unwind in (unwind = true, $( $key = $value ),* ) {
+                    $crate::comptime_if::comptime_if! {
+                        if unwind where (unwind = true, $( $key = $value ),* ) {
                             $crate::input::__bridge::initialize_plugin_c_unwind::<$struct>(version)
                         } else {
                             $crate::input::__bridge::initialize_plugin_c::<$struct>(version)
@@ -814,8 +814,8 @@ macro_rules! register_input_plugin {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn UninitializePlugin() {
                 unsafe {
-                    $crate::__macro_if! {
-                        if unwind in (unwind = true, $( $key = $value ),* ) {
+                    $crate::comptime_if::comptime_if! {
+                        if unwind where (unwind = true, $( $key = $value ),* ) {
                             $crate::input::__bridge::uninitialize_plugin_c_unwind::<$struct>()
                         } else {
                             $crate::input::__bridge::uninitialize_plugin::<$struct>()
@@ -827,8 +827,8 @@ macro_rules! register_input_plugin {
             #[unsafe(no_mangle)]
             unsafe extern "C" fn GetInputPluginTable()
             -> *mut aviutl2::sys::input2::INPUT_PLUGIN_TABLE {
-                $crate::__macro_if! {
-                    if unwind in (unwind = true, $( $key = $value ),* ) {
+                $crate::comptime_if::comptime_if! {
+                    if unwind where (unwind = true, $( $key = $value ),* ) {
                         unsafe { $crate::input::__bridge::create_table_unwind::<$struct>() }
                     } else {
                         unsafe { $crate::input::__bridge::create_table::<$struct>() }