+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-（なし）
+- module: i32・f64以外にもFromScriptModuleParamを実装
+- ログを自動で折り返すように
 
 ### デモプラグイン